move tx rebalance validators
