+++ x/stakeibc/client/cli/tx.go
@@ -288,3 +288,36 @@ func CmdClaimUndelegatedTokens() *cobra.Command {
 
 	return cmd
 }
+
+func CmdRebalanceValidators() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "rebalance-validators [host-zone] [num-to-rebalance]",
+		Short: "Broadcast message rebalanceValidators",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argHostZone := args[0]
+			argNumValidators, err := strconv.ParseUint(args[1], 10, 64)
+			if err != nil {
+				return err
+			}
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgRebalanceValidators(
+				clientCtx.GetFromAddress().String(),
+				argHostZone,
+				argNumValidators,
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
