+++ x/stakeibc/client/cli/tx_rebalance_validators.go
@@ -2,46 +2,6 @@ package cli
 
 import (
 	"strconv"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	"github.com/spf13/cobra"
-
-	"github.com/Stride-Labs/stride/v22/x/stakeibc/types"
 )
 
 var _ = strconv.Itoa(0)
-
-func CmdRebalanceValidators() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "rebalance-validators [host-zone] [num-to-rebalance]",
-		Short: "Broadcast message rebalanceValidators",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			argHostZone := args[0]
-			argNumValidators, err := strconv.ParseUint(args[1], 10, 64)
-			if err != nil {
-				return err
-			}
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			msg := types.NewMsgRebalanceValidators(
-				clientCtx.GetFromAddress().String(),
-				argHostZone,
-				argNumValidators,
-			)
-			if err := msg.ValidateBasic(); err != nil {
-				return err
-			}
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
